requirements.txt
aiogram>=3.17.0
langchain>=0.3.14
langchain-community>=0.3.19
langchain-openai==0.2.3
pandas>=2.2.3
openai>=1.59.7
python-dotenv>=1.0.1